(vim) open panes to the right, enable mouse
Mouse is useful for pairing, since people tend to reach for mouse
scroll. Unfortunately it breaks some of the text selection/copying
features in Tmux, since if I'm in Vim, it will default to a visual
selection rather than a Tmux text selection, but I feel like it's a
worthwhile tradeoff, and easily worked around using option+mouse select.

Splitright is because normally when I'm opening a new pane it's as a
secondary view, so I don't want it to open in what I think of as the
'main' view (the left).