WIP Attempt to switch rspec command
When running through spring we want to make sure we're running bin/rspec
so that we get the benefits of spring, but when that doesn't exist we
need to fall back to bundle exec rspec.

I guess this is modelled after some `rails_command` (?) function defined
by oh-my-zsh? I should look into that more

I haven't been using this: I think it had some issues, but I don't
remember what.