(git) use the diff-so-fancy pager
https://github.com/so-fancy/diff-so-fancy

This provides a number of benefits, including showing which segment of a
line changed.

The `less` command means that the diff view starts from the top of the
diff rather than just showing the end and making me scroll.

stripLeadingSymbols is false because otherwise in merge conflict diffs
it's hard to see which line came from which commit.