Battery function returns null if no battery
This was producing empty output on my mac mini, which was breaking my
command prompt.

The solution is to check whether there's actually a battery and only
include it in the output if there is one.

Note that we need to do some shenanigans to handle the % character: this
is a special character so we need to escape it by using %%, but the
challenge is that the battery string already includes a %, so we need to
add just one more.

Note that the battery function accepts an argument which adds this
second %, since it's only needed when the function is being used inside
a command prompt (or other interpreted string) - otherwise we want to
display it with just one % character.