(vim) run RSpec outside of Tmux
Inline images in spec output are nice, but for Reasons, they don't work
inside Tmux and are apparently unlikely to ever.

I experimented with the tmux integration but kind of hated it, so
instead let's open a new iTerm tab outside of Tmux and run specs there.
This took AGES to work out how to do.

Needs an autocmd so that we calculate the buffer vars on every page
load. This is probably overkill (could just make it global vars
evaluated on every page load?) but felt like the smallest change.