(iterm) Default to darkmode
This is very annoying: when I toggle between dark and light it changes
these colour values in the config, so when I close iterm in a different
mode it detects this as a config change and asks if I want to write the
changes to the config file.

I guess I'm defaulting more to darkmode, so let's make that the default.